Break down tests/eth2/core/beacon/helpers.py
