+++ tests/eth2/core/beacon/state_machines/forks/test_serenity_block_processing.py
@@ -38,8 +38,8 @@ from eth2.beacon.tools.builder.proposer import (
     _generate_randao_reveal,
 )
 
-from tests.eth2.beacon.helpers import (
-    mock_validator_record,
+from eth2.beacon.tools.builder.initializer import (
+    mock_validator,
 )
 
 
@@ -51,7 +51,7 @@ def test_randao_processing(sample_beacon_block_params,
     proposer_pubkey, proposer_privkey = first(keymap.items())
     state = SerenityBeaconState(**sample_beacon_state_params).copy(
         validator_registry=tuple(
-            mock_validator_record(proposer_pubkey, config)
+            mock_validator(proposer_pubkey, config)
             for _ in range(config.TARGET_COMMITTEE_SIZE)
         ),
         validator_balances=(config.MAX_DEPOSIT_AMOUNT,) * config.TARGET_COMMITTEE_SIZE,
@@ -101,7 +101,7 @@ def test_randao_processing_validates_randao_reveal(sample_beacon_block_params,
     proposer_pubkey, proposer_privkey = first(keymap.items())
     state = SerenityBeaconState(**sample_beacon_state_params).copy(
         validator_registry=tuple(
-            mock_validator_record(proposer_pubkey, config)
+            mock_validator(proposer_pubkey, config)
             for _ in range(config.TARGET_COMMITTEE_SIZE)
         ),
         validator_balances=(config.MAX_DEPOSIT_AMOUNT,) * config.TARGET_COMMITTEE_SIZE,
