+++ tests/eth2/core/beacon/state_machines/forks/test_serenity_block_validation.py
@@ -42,7 +42,7 @@ from eth2.beacon.types.forks import Fork
 from eth2.beacon.types.slashable_attestations import SlashableAttestation
 from eth2.beacon.types.states import BeaconState
 
-from tests.eth2.beacon.helpers import mock_validator_record
+from eth2.beacon.tools.builder.initializer import mock_validator
 
 
 @pytest.mark.parametrize(
@@ -97,7 +97,7 @@ def test_validate_proposer_signature(
 
     state = BeaconState(**sample_beacon_state_params).copy(
         validator_registry=tuple(
-            mock_validator_record(proposer_pubkey, config)
+            mock_validator(proposer_pubkey, config)
             for _ in range(10)
         ),
         validator_balances=(max_deposit_amount,) * 10,
